% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_tc_add_sample_meta.R
\name{tc_add_sample_meta}
\alias{tc_add_sample_meta}
\title{Add Sample-Level Metadata to a Seurat Object}
\usage{
tc_add_sample_meta(
  object,
  meta_df,
  join_by = "orig.ident",
  factor_levels = NULL,
  set_ident = NULL
)
}
\arguments{
\item{object}{A Seurat object.}

\item{meta_df}{A data.frame containing the metadata.}

\item{join_by}{A character string specifying the column name to match.}

\item{factor_levels}{A named list specifying factor levels.}

\item{set_ident}{A character string to set as Idents.}
}
\value{
A Seurat object with updated metadata.
}
\description{
This function merges a data frame containing sample-level clinical or experimental
metadata into a Seurat object's cell-level metadata.
}
