% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tc_qc_filter.R
\name{tc_qc_filter}
\alias{tc_qc_filter}
\title{Quality Control and Filtering of a Seurat Object}
\usage{
tc_qc_filter(
  object,
  mt_pattern = "^MT-",
  min_features = 300,
  max_features = 7500,
  min_counts = 500,
  max_counts = Inf,
  max_mt = 20
)
}
\arguments{
\item{object}{A Seurat object.}

\item{mt_pattern}{Regex pattern for mitochondrial genes. Default is \code{"^MT-"}.}

\item{min_features}{Min genes per cell. Default is 300.}

\item{max_features}{Max genes per cell. Default is 7500.}

\item{min_counts}{Min total counts (UMI) per cell. Default is 500.}

\item{max_counts}{Max total counts (UMI) per cell. Default is Inf (no upper limit).}

\item{max_mt}{Max percentage of mitochondrial genes. Default is 20.}
}
\value{
A filtered \code{Seurat} object.
}
\description{
This function calculates mitochondrial gene percentage and filters cells based on
features (genes), total counts (UMIs), and mitochondrial content.
}
\examples{
\dontrun{
sce.filt <- tc_qc_filter(
  object = sce.all,
  min_features = 300,
  max_features = 7500,
  min_counts = 500,
  max_mt = 20
)
}
}
